Split referrer tests into StatsRepositoryReferrersTest to fix LargeClass
StatsRepositoryTest tripped detekt's LargeClass rule after the referrer
tests were added. Extract them into a dedicated test class instead of
baselining, keeping the referrer tests grouped in one place.
